Improve: add useAdminCoupons composable
